- saner logging code - ui resources are created using the standard Toolkit convention
