+++ python/tank_vendor/shotgun_authentication/resources/build_resources.sh
@@ -12,6 +12,7 @@
 
 # The path to output all built .py files to: 
 UI_PYTHON_PATH=../ui
+PYTHON_BASE="/Applications/Shotgun.app/Contents/Frameworks/Python"
 
 
 # Helper functions to build UI files
@@ -26,11 +27,11 @@ function build_qt {
 }
 
 function build_ui {
-    build_qt "pyside-uic --from-imports" "$1.ui" "../ui/ui_$1"
+    build_qt "${PYTHON_BASE}/bin/python ${PYTHON_BASE}/bin/pyside-uic --from-imports" "$1.ui" "../ui/$1"
 }
 
 function build_res {
-    build_qt "pyside-rcc" "$1.qrc" "../ui/$1_rc"
+    build_qt "${PYTHON_BASE}/bin/pyside-rcc" "$1.qrc" "../ui/$1_rc"
 }
