+++ python/tank_vendor/shotgun_authentication/ui/login_dialog.py
@@ -1,169 +1,129 @@
-# Copyright (c) 2013 Shotgun Software Inc.
-#
-# CONFIDENTIAL AND PROPRIETARY
-#
-# This work is provided "AS IS" and subject to the Shotgun Pipeline Toolkit
-# Source Code License included in this distribution package. See LICENSE.
-# By accessing, using, copying or modifying this work you indicate your
-# agreement to the Shotgun Pipeline Toolkit Source Code License. All rights
-# not expressly granted therein are reserved by Shotgun Software Inc.
+# -*- coding: utf-8 -*-
 
-"""
-Login dialog for authenticating to a Shotgun server.
-"""
+# Form implementation generated from reading ui file 'login_dialog.ui'
+#
+#      by: pyside-uic 0.2.15 running on PySide 1.2.1
+#
+# WARNING! All changes made in this file will be lost!
 
+from .qt_abstraction import QtCore, QtGui
+
+class Ui_LoginDialog(object):
+    def setupUi(self, LoginDialog):
+        LoginDialog.setObjectName("LoginDialog")
+        LoginDialog.setWindowModality(QtCore.Qt.NonModal)
+        LoginDialog.resize(374, 324)
+        LoginDialog.setStyleSheet("QWidget\n"
+"{\n"
+"    background-color:  rgb(36, 39, 42);\n"
+"    color: rgb(192, 193, 195);\n"
+"    selection-background-color: rgb(168, 123, 43);\n"
+"    selection-color: rgb(230, 230, 230);\n"
+"    font-size: 11px;\n"
+"}\n"
+"\n"
+"QPushButton\n"
+"{\n"
+"    background-color: transparent;\n"
+"    border-radius: 2px;\n"
+"    padding: 8px;\n"
+"    padding-left: 15px;\n"
+"    padding-right: 15px;\n"
+"}\n"
+"\n"
+"QLineEdit\n"
+"{\n"
+"    background-color: rgb(29, 31, 34);\n"
+"    border: 1px solid rgb(54, 60, 66);\n"
+"    border-radius: 2px;\n"
+"    padding: 5px;\n"
+"    font-size: 12px;\n"
+"}\n"
+"\n"
+"QLineEdit:focus\n"
+"{\n"
+"    border: 1px solid rgb(48, 167, 227);\n"
+"\n"
+"}")
+        LoginDialog.setModal(True)
+        self.verticalLayout_2 = QtGui.QVBoxLayout(LoginDialog)
+        self.verticalLayout_2.setContentsMargins(20, 20, 20, 20)
+        self.verticalLayout_2.setObjectName("verticalLayout_2")
+        self.horizontalLayout = QtGui.QHBoxLayout()
+        self.horizontalLayout.setSpacing(0)
+        self.horizontalLayout.setSizeConstraint(QtGui.QLayout.SetMinAndMaxSize)
+        self.horizontalLayout.setObjectName("horizontalLayout")
+        self.logo = AspectPreservingLabel(LoginDialog)
+        sizePolicy = QtGui.QSizePolicy(QtGui.QSizePolicy.Preferred, QtGui.QSizePolicy.Preferred)
+        sizePolicy.setHorizontalStretch(0)
+        sizePolicy.setVerticalStretch(0)
+        sizePolicy.setHeightForWidth(self.logo.sizePolicy().hasHeightForWidth())
+        self.logo.setSizePolicy(sizePolicy)
+        self.logo.setMaximumSize(QtCore.QSize(250, 150))
+        self.logo.setText("")
+        self.logo.setAlignment(QtCore.Qt.AlignCenter)
+        self.logo.setTextInteractionFlags(QtCore.Qt.NoTextInteraction)
+        self.logo.setObjectName("logo")
+        self.horizontalLayout.addWidget(self.logo)
+        self.verticalLayout_2.addLayout(self.horizontalLayout)
+        self.site = QtGui.QLineEdit(LoginDialog)
+        self.site.setMinimumSize(QtCore.QSize(308, 0))
+        self.site.setObjectName("site")
+        self.verticalLayout_2.addWidget(self.site)
+        self.login = QtGui.QLineEdit(LoginDialog)
+        self.login.setMinimumSize(QtCore.QSize(308, 0))
+        self.login.setObjectName("login")
+        self.verticalLayout_2.addWidget(self.login)
+        self.password = QtGui.QLineEdit(LoginDialog)
+        self.password.setMinimumSize(QtCore.QSize(308, 0))
+        self.password.setEchoMode(QtGui.QLineEdit.Password)
+        self.password.setObjectName("password")
+        self.verticalLayout_2.addWidget(self.password)
+        self.message = QtGui.QLabel(LoginDialog)
+        sizePolicy = QtGui.QSizePolicy(QtGui.QSizePolicy.Preferred, QtGui.QSizePolicy.Fixed)
+        sizePolicy.setHorizontalStretch(0)
+        sizePolicy.setVerticalStretch(0)
+        sizePolicy.setHeightForWidth(self.message.sizePolicy().hasHeightForWidth())
+        self.message.setSizePolicy(sizePolicy)
+        self.message.setText("")
+        self.message.setAlignment(QtCore.Qt.AlignLeading|QtCore.Qt.AlignLeft|QtCore.Qt.AlignVCenter)
+        self.message.setWordWrap(True)
+        self.message.setMargin(4)
+        self.message.setObjectName("message")
+        self.verticalLayout_2.addWidget(self.message)
+        self.button_layout = QtGui.QHBoxLayout()
+        self.button_layout.setSpacing(10)
+        self.button_layout.setContentsMargins(5, -1, -1, -1)
+        self.button_layout.setObjectName("button_layout")
+        spacerItem = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
+        self.button_layout.addItem(spacerItem)
+        self.cancel = QtGui.QPushButton(LoginDialog)
+        self.cancel.setStyleSheet("")
+        self.cancel.setAutoDefault(False)
+        self.cancel.setFlat(True)
+        self.cancel.setObjectName("cancel")
+        self.button_layout.addWidget(self.cancel)
+        self.sign_in = QtGui.QPushButton(LoginDialog)
+        self.sign_in.setStyleSheet("color: rgb(248, 248, 248);\n"
+"background-color: rgb(35, 165, 225);")
+        self.sign_in.setAutoDefault(False)
+        self.sign_in.setDefault(True)
+        self.sign_in.setFlat(True)
+        self.sign_in.setObjectName("sign_in")
+        self.button_layout.addWidget(self.sign_in)
+        self.button_layout.setStretch(0, 1)
+        self.verticalLayout_2.addLayout(self.button_layout)
+        self.verticalLayout_2.setStretch(0, 1)
+
+        self.retranslateUi(LoginDialog)
+        QtCore.QMetaObject.connectSlotsByName(LoginDialog)
+
+    def retranslateUi(self, LoginDialog):
+        self.site.setPlaceholderText(QtGui.QApplication.translate("LoginDialog", "example.shotgunstudio.com", None, QtGui.QApplication.UnicodeUTF8))
+        self.login.setPlaceholderText(QtGui.QApplication.translate("LoginDialog", "login", None, QtGui.QApplication.UnicodeUTF8))
+        self.password.setPlaceholderText(QtGui.QApplication.translate("LoginDialog", "password", None, QtGui.QApplication.UnicodeUTF8))
+        self.cancel.setText(QtGui.QApplication.translate("LoginDialog", "Cancel", None, QtGui.QApplication.UnicodeUTF8))
+        self.sign_in.setText(QtGui.QApplication.translate("LoginDialog", "Sign In", None, QtGui.QApplication.UnicodeUTF8))
+
+from .aspect_preserving_label import AspectPreservingLabel
 from . import resources_rc
-from . import ui_login_dialog
-from .. import session_cache
-from ..errors import AuthenticationError
-from .qt_abstraction import QtGui, QtCore
-
-
-class LoginDialog(QtGui.QDialog):
-    """
-    Dialog for getting user crendentials.
-    """
-    def __init__(self, title, is_session_renewal, hostname="", login="", fixed_host=False, http_proxy=None, parent=None):
-        """
-        Constructs a dialog.
-
-        :param title: Title of this dialog.
-        :param is_session_renewal: Boolean indicating if we are renewing a session or authenticating a user from scratch.
-        :param hostname: The string to populate the site field with. Defaults to "".
-        :param login: The string to populate the login field with. Defaults to "".
-        :param fixed_host: Indicates if the hostname can be changed. Defaults to False.
-        :param http_proxy: The proxy server to use when testing authentication. Defaults to None.
-        :param parent: The Qt parent for the dialog (defaults to None)
-        """
-        QtGui.QDialog.__init__(self, parent)
-
-        # setup the gui
-        self.ui = ui_login_dialog.Ui_LoginDialog()
-        self.ui.setupUi(self)
-
-        # Set the title
-        self.setWindowTitle(title)
-
-        # Assign credentials
-        self._http_proxy = http_proxy
-        self.ui.site.setText(hostname)
-        self.ui.login.setText(login)
-
-        # default focus
-        if self.ui.site.text():
-            if self.ui.login.text():
-                self.ui.password.setFocus()
-            else:
-                self.ui.login.setFocus()
-        else:
-            self.ui.site.setFocus()
-
-        # set the logo
-        self.ui.logo.setPixmap(QtGui.QPixmap(":/shotgun_authentication/shotgun_logo_light_medium.png"))
-
-        # Disable keyboard input in the site and login boxes if we are simply renewing the session.
-        self.ui.site.setReadOnly(is_session_renewal or fixed_host)
-        self.ui.login.setReadOnly(is_session_renewal)
-
-        if is_session_renewal:
-            self._set_error_message("Your session has expired. Please enter your Shotgun password.")
-        else:
-            self._set_message("Please enter your Shotgun credentials.")
-
-        # hook up signals
-        self.ui.sign_in.clicked.connect(self._ok_pressed)
-        self.ui.cancel.clicked.connect(self.reject)
-
-    def _set_message(self, message):
-        """
-        Set the message in the dialog.
-        :param message: Message to display in the dialog.
-        """
-        if not message:
-            self.ui.message.hide()
-        else:
-            self.ui.message.setText(message)
-            self.ui.message.show()
-
-    def show(self):
-        """
-        Shows the ui.
-        """
-        QtGui.QDialog.show(self)
-        self.activateWindow()
-        self.raise_()
-
-    def exec_(self):
-        """
-        Displays the window modally.
-        """
-        self.activateWindow()
-        self.raise_()
-        # the trick of activating + raising does not seem to be enough for
-        # modal dialogs. So force put them on top as well.
-        self.setWindowFlags(QtCore.Qt.WindowStaysOnTopHint | self.windowFlags())
-        return QtGui.QDialog.exec_(self)
-
-    def result(self):
-        """
-        Displays a modal dialog asking for the credentials.
-        :returns: A tuple of (hostname, username and session token) string if the user authenticated
-                  None if the user cancelled.
-        """
-        if self.exec_() == QtGui.QDialog.Accepted:
-            return (self.ui.site.text().encode("utf-8"),
-                    self.ui.login.text().encode("utf-8"),
-                    self._new_session_token)
-        else:
-            return None
-
-    def _set_error_message(self, message):
-        """
-        Set the error message in the dialog.
-        :param message: Message to display in red in the dialog.
-        """
-        self.ui.message.setText("<font style='color: rgb(252, 98, 70);'>%s</font>" % message)
-
-    def _ok_pressed(self):
-        """
-        validate the values, accepting if login is successful and display an error message if not.
-        """
-        # pull values from the gui
-        site = self.ui.site.text()
-        login = self.ui.login.text()
-        password = self.ui.password.text()
-
-        if len(site) == 0:
-            self._set_error_message("Please enter the address of the site to connect to.")
-            return
-        if len(login) == 0:
-            self._set_error_message("Please enter your login name.")
-            return
-        if len(password) == 0:
-            self._set_error_message("Please enter your password.")
-            return
-
-        # if not protocol specified assume https
-        if len(site.split("://")) == 1:
-            site = "https://%s" % site
-            self.ui.site.setText(site)
-
-        try:
-            # set the wait cursor
-            QtGui.QApplication.setOverrideCursor(QtCore.Qt.WaitCursor)
-            QtGui.QApplication.processEvents()
-
-            # try and authenticate
-            self._new_session_token = session_cache.generate_session_token(
-                site, login, password, self._http_proxy
-            )
-        except AuthenticationError, e:
-            # authentication did not succeed
-            self._set_error_message(e[0])
-            self.ui.message.show()
-            return
-        finally:
-            # restore the cursor
-            QtGui.QApplication.restoreOverrideCursor()
-            QtGui.QApplication.processEvents()        # dialog is done
-        self.accept()
